requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
pip==25.1.1
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic[email]==2.11.7
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
rsa==4.9.1
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-extensions==4.14.0
typing-inspection==0.4.1
uvicorn==0.35.0
